copy patch file properly

+++ scripts/copy-modules.js
@@ -2,10 +2,13 @@
 var fs = require('fs');
 var path = require('path');
 
-var srcDir = path.resolve(__dirname, './../pkgs/react-native-audio-record/ios');
+var srcDir = path.resolve(
+  __dirname,
+  './../../pkgs/react-native-audio-record/ios',
+);
 var destDir = path.resolve(
   __dirname,
-  './node_modules/react-native-audio-record/ios',
+  './../node_modules/react-native-audio-record/ios',
 );
 
 fs.readdir(srcDir, function (err, files) {
@@ -17,6 +20,8 @@ fs.readdir(srcDir, function (err, files) {
     var srcFile = path.join(srcDir, file);
     var destFile = path.join(destDir, file);
 
-    fs.copyFileSync(srcFile, destFile);
+    if (fs.statSync(srcFile).isFile()) {
+      fs.copyFileSync(srcFile, destFile);
+    }
   });
 });
